test/tmux/requirements.txt
asciinema==v2.0.2
pytest-tmux
docker
